Profile page bug fixes
profile page was trying to render the page before resources were received from the server, still figuring out why rendering that caused a crash but for now it loads things in the correct order.

The user controller on the server side now verifies user objects before sending them using.

userUtils now has a verify object function that just makes sure the object fed to it is a compete userObject

some of the controllers have been modified to utilize these changes