commonNames for ingredients
ingredients.list api has been renamed to ingredients.find. This is to standardize the naming conventions of the api's

a new table now exists inside the postgres database called "common_food_names". This will have a list of all common names for food_ids so if the user is looking up something with a name that may be hard to find in the database, the user can search for the common name and the ingredient.find api will know to prioritize the ingredient with the common name.

other then that, just some visual changes have been made.